logro convertido en átomo
