+++ src/components/atoms/index.js
@@ -3,5 +3,5 @@ import TextInputEditForm from './text-inputEditForm';
 import Button from './button';
 import FormHeader from './form-header';
 import List from './list';
-
-export { TextInput, Button, FormHeader, TextInputEditForm, List };
+import Logro from './logro';
+export { TextInput, Button, FormHeader, TextInputEditForm, List, Logro };
