+++ src/components/atoms/logro/index.js
@@ -1,0 +1,3 @@
+import View from './view';
+
+export default View;
