+++ src/components/pages/meLogros/styled.js
@@ -36,32 +36,6 @@ export const RowLogros = styled.div`
   max-width: 100%;
 `;
 
-export const ContainerLogros = styled.div`
-  align-items: center;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  margin: 0 auto;
-  text-align: center;
-  width: 230px;
-
-  @media screen and (max-width: 968px) {
-    width: 33%;
-  }
-
-  @media screen and (max-width: 767px) {
-    width: 50%;
-  }
-`;
-
-export const LogroImg = styled.img`
-  background-color: ${colors.lightGrey};
-  border-radius: 50%;
-  height: 100px;
-  margin: 0.5rem;
-  width: 100px;
-`;
-
 export const TextWrapper = styled.div`
   align-items: center;
   justify-content: center;
@@ -148,25 +122,5 @@ export const ButtonDelete = styled.button`
   @media screen and (max-width: 960px) {
     display: block;
     margin: 1rem auto;
-  }
-`;
-
-export const NameText = styled.p`
-  color: ${colors.darkGrey};
-  font-size: 1rem;
-  font-weight: bold;
-  padding: 1rem 1rem 0 1rem;
-`;
-
-export const DescriptionText = styled.p`
-  color: ${colors.darkGrey};
-  font-size: 1rem;
-  font-weight: 300;
-  line-height: 1.5;
-  padding: 0.5rem 1rem 1rem 1rem;
-
-  @media screen and (max-width: 768px) {
-    font-size: 0.813rem;
-    padding: 0.5rem 0rem;
   }
 `;
